+++ app/src/main/java/com/samsung/monimo/UI/product/ProductInterestFragment.kt
@@ -1,0 +1,19 @@
+package com.samsung.monimo.UI.product
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.samsung.monimo.R
+
+class ProductInterestFragment : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_product_interest, container, false)
+    }
+}
