+++ app/src/main/java/com/samsung/monimo/UI/setting/OnboardingFragment.kt
@@ -1,4 +1,4 @@
-package com.samsung.monimo.UI
+package com.samsung.monimo.UI.setting
 
 import android.os.Bundle
 import android.view.LayoutInflater
