+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="com.samsung.monimo.UI.OnboardingFragment"
+        android:name="com.samsung.monimo.UI.setting.OnboardingFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
 </LinearLayout>
