+++ app/src/main/res/layout/bottom_sheet_search.xml
@@ -42,10 +42,10 @@
         android:layout_marginTop="15dp"
         android:background="@drawable/searchview_background"
         android:drawableLeft="@drawable/ic_search_resize"
+        android:drawablePadding="20dp"
         android:ems="10"
         android:inputType="text"
         android:padding="15dp"
-        tools:drawablePadding="20dp"
         tools:inputType="textPersonName" />
 
     <com.google.android.material.divider.MaterialDivider
