+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.OnboardingFragment" >
+    tools:context=".UI.setting.OnboardingFragment" >
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
@@ -22,7 +22,7 @@
         android:id="@+id/textViewOnboarding"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="50dp"
+        android:layout_marginTop="100dp"
         android:fontFamily="@font/pretendard_semi_bold"
         android:text="내 집 마련은 언제?"
         android:textAlignment="center"
