+++ app/src/main/res/layout/fragment_product_earning.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".UI.ProductEarningFragment">
+    tools:context=".UI.product.ProductEarningFragment">
 
     <LinearLayout
         android:id="@+id/linearLayoutTop"
