+++ app/src/main/res/layout/fragment_product_interest.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".UI.ProductInterestFragment">
+    tools:context=".UI.product.ProductInterestFragment">
 
     <LinearLayout
         android:id="@+id/linearLayoutTop"
