+++ app/src/main/res/layout/fragment_search_house.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".UI.SearchHouseFragment">
+    tools:context=".UI.setting.SearchHouseFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
