+++ app/src/main/res/layout/fragment_search_product.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".UI.SearchProductFragment">
+    tools:context=".UI.product.SearchProductFragment">
 
     <TextView
         android:id="@+id/textViewSearch"
