+++ app/src/main/res/layout/fragment_setting_period.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".UI.SettingPeriodFragment">
+    tools:context=".UI.setting.SettingPeriodFragment">
 
     <com.google.android.material.appbar.MaterialToolbar
         android:id="@+id/toolbar"
